pyproject.toml
[tool.poetry]
name = "my-fastapi-app"
version = "0.1.0"
description = "Example of FastAPI with Opentelemetry and Azure Monitor"
authors = ["Alexander Bij <Alexander.Bij@xebia.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.11"}
azure-monitor-opentelemetry = "^1.6.5"
opentelemetry-instrumentation-fastapi = "^0.52b1"
opentelemetry-instrumentation-requests = "^0.52b1"
opentelemetry-instrumentation-httpx = "^0.52b1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
